pyproject.toml
[project]
name = "print-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jonathan Starr", email = "github@jstarr.me" }
]
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
    "pycups>=2.0.4",
    "python-barcode>=0.16.1",
    "pyudev>=0.24.4",
]

[project.scripts]
print-server = "print_server.__main__:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.8.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.3.0",
    "types-pillow>=10.2.0.20240822",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
